requirements.txt
requests
python-dotenv
pandas
sqlalchemy
schedule